pyproject.toml
[tool.poetry]
name = "cursovaya"
version = "0.1.0"
description = ""
authors = ["Stepan <stepstepan2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.2"
xlrd = "^2.0.1"
xlwt = "^1.3.0"
openpyxl = "^3.1.5"
simplejson = "^3.19.2"
load-dotenv = "^0.1.0"
requests = "^2.32.3"


[tool.poetry.group.lint.dependencies]
mypy = "^1.10.1"
black = "^24.4.2"
isort = "^5.13.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.0"
pytest = "^8.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

[tool.black]
# Максимальная длина строки
line-length = 119
# Файлы, которые не нужно форматировать
exclude = '''
(
  /(
      \.eggs         # Исключить несколько общих каталогов
    | \.git          # в корне проекта
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py           # Также отдельно исключить файл с именем foo.py
                     # в корне проекта
)
'''

[tool.isort]
# максимальная длина строки
line_length = 119
